# RacingGuru/internal/controller/handlers/assets/swagger/swagger.yaml
basePath: /
definitions:
  handlers.ChangePasswordRequest:
    properties:
      password:
        example: newSecret123
        type: string
    type: object
  handlers.AuthUserResponse:
    properties:
      email:
        example: user@example.com
        type: string
      name:
        example: Max Verstappen
        type: string
      role:
        allOf:
        - $ref: '#/definitions/models.Role'
        enum:
        - admin
        - user
        example: user
    type: object
  handlers.GenerateAdminSecretResponse:
    properties:
      secret:
        example: WQ3mYvB1d8jK2s4Qn7LpT0xZc5RaE9Uf
        type: string
    type: object
  handlers.DriverStatsResponse:
    properties:
      driver_stats:
        $ref: '#/definitions/models.DriverStats'
    type: object
  handlers.ErrorResponse:
    properties:
      error:
        properties:
          code:
            type: string
          message:
            type: string
        type: object
    type: object
  handlers.LoginRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: secret123
        type: string
    type: object
  handlers.RegisterRequest:
    properties:
      email:
        example: user@example.com
        type: string
      name:
        example: Max Verstappen
        type: string
      password:
        example: secret123
        type: string
      secret:
        example: WQ3mYvB1d8jK2s4Qn7LpT0xZc5RaE9Uf
        type: string
    type: object
  handlers.TeamStatsResponse:
    properties:
      team_stats:
        $ref: '#/definitions/models.TeamStats'
    type: object
  handlers.UpdateUserRoleRequest:
    properties:
      id:
        example: 77777777-7777-7777-7777-777777777777
        type: string
      role:
        allOf:
        - $ref: '#/definitions/models.Role'
        enum:
        - admin
        - user
        example: admin
    type: object
  handlers.tokenS:
    properties:
      token:
        type: string
    type: object
  models.CarParticipant:
    properties:
      car_id:
        type: string
      drivers:
        items:
          $ref: '#/definitions/models.Driver'
        type: array
      id:
        type: string
      number:
        type: string
      team_id:
        type: string
    type: object
  models.ConditionOperator:
    enum:
    - gt
    - lt
    - gte
    - lte
    - eq
    type: string
    x-enum-varnames:
    - ConditionOperatorGT
    - ConditionOperatorLT
    - ConditionOperatorGTE
    - ConditionOperatorLTE
    - ConditionOperatorEQ
  models.Driver:
    properties:
      birthday:
        type: string
      id:
        type: string
      name:
        type: string
      nationality:
        type: string
    type: object
  models.DriverStats:
    properties:
      driver:
        $ref: '#/definitions/models.Driver'
      stats:
        $ref: '#/definitions/models.Stats'
    type: object
  models.MatrixDrivers:
    properties:
      condition_specs:
        items:
          items:
            $ref: '#/definitions/models.SudokuCondition'
          type: array
        type: array
      field:
        items:
          items:
            items:
              $ref: '#/definitions/models.Driver'
            type: array
          type: array
        type: array
    type: object
  models.MatrixTeams:
    properties:
      condition_specs:
        items:
          items:
            $ref: '#/definitions/models.SudokuCondition'
          type: array
        type: array
      field:
        items:
          items:
            items:
              $ref: '#/definitions/models.Team'
            type: array
          type: array
        type: array
    type: object
  models.Race:
    properties:
      championship_id:
        type: string
      date:
        type: string
      duration:
        type: integer
      id:
        type: string
      name:
        type: string
      track:
        $ref: '#/definitions/models.Track'
      type:
        type: integer
    type: object
  models.RaceResult:
    properties:
      car_participant_id:
        type: string
      finish_pos:
        type: integer
      qualifying_pos:
        type: integer
      race_id:
        type: string
    type: object
  models.Role:
    enum:
    - admin
    - user
    type: string
    x-enum-varnames:
    - RoleAdmin
    - RoleUser
  models.Stats:
    properties:
      best_championship_position:
        type: integer
      best_finish:
        type: integer
      best_qualifying:
        type: integer
      championships_wins:
        type: integer
      total_podiums:
        type: integer
      total_points:
        type: integer
      total_poles:
        type: integer
      total_races:
        type: integer
      total_wins:
        type: integer
    type: object
  models.SudokuCondition:
    properties:
      field:
        type: string
      label:
        type: string
      op:
        $ref: '#/definitions/models.ConditionOperator'
      value:
        type: integer
    type: object
  models.Team:
    properties:
      country:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  models.TeamStats:
    properties:
      stats:
        $ref: '#/definitions/models.Stats'
      team:
        $ref: '#/definitions/models.Team'
    type: object
  models.Track:
    properties:
      country:
        type: string
      id:
        type: string
      length:
        type: integer
      name:
        type: string
      turns:
        type: integer
    type: object
  models.User:
    properties:
      email:
        type: string
      favourite_driers:
        items:
          $ref: '#/definitions/models.Driver'
        type: array
      favourite_teams:
        items:
          $ref: '#/definitions/models.Team'
        type: array
      id:
        type: string
      name:
        type: string
      password:
        type: string
      role:
        $ref: '#/definitions/models.Role'
    type: object
info:
  contact: {}
  description: HTTP API for authentication and racing statistics.
  title: RacingGuru API
  version: 0.1.0
paths:
...
  /admin/car-participants:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Car participant payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CarParticipant'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CarParticipant'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Update car participant
      tags:
      - admin
    post:
      consumes:
      - application/json
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Car participant payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CarParticipant'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.CarParticipant'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Create car participant
      tags:
      - admin
  /admin/car-participants/drivers:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Car participant drivers payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CarParticipant'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CarParticipant'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Update car participant drivers
      tags:
      - admin
  /admin/drivers:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Driver payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Driver'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Driver'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Update driver
      tags:
      - admin
    post:
      consumes:
      - application/json
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Driver payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Driver'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Driver'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Create driver
      tags:
      - admin
  /admin/race-results:
    post:
      consumes:
      - application/json
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Race result payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.RaceResult'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.RaceResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Upsert race result
      tags:
      - admin
  /admin/races:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Race payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Race'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Race'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Update race
      tags:
      - admin
    post:
      consumes:
      - application/json
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Race payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Race'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Race'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Create race
      tags:
      - admin
  /admin/teams:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Team payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Team'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Team'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Update team
      tags:
      - admin
    post:
      consumes:
      - application/json
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Team payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Team'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Team'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Create team
      tags:
      - admin
  /admin/tracks:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Track payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Track'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Track'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Update track
      tags:
      - admin
    post:
      consumes:
      - application/json
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Track payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Track'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Track'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Create track
      tags:
      - admin
  /admin/users/role:
    patch:
      consumes:
      - application/json
      description: Changes another user's role. Only users with role `admin` can call
        this endpoint.
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User role update payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateUserRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Update user role
      tags:
      - admin
  /change-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: New password payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Change user password
      tags:
      - auth
...
  /login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.tokenS'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: User login
      tags:
      - auth
  /register:
    post:
      consumes:
      - application/json
      description: Registers a new user. If a valid one-time `secret` is provided, registers the first admin instead.
      parameters:
      - description: Registration payload. Optional secret promotes the first user to admin
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Registered user
          schema:
            $ref: '#/definitions/handlers.AuthUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: User registration
      tags:
      - auth
  /setup/admin-secret:
    post:
      consumes:
      - application/json
      description: Generates a one-time secret that can be passed in `/register` to create the first admin user.
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.GenerateAdminSecretResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Generate one-time admin secret
      tags:
      - auth
  /stats/driver:
    get:
      parameters:
      - description: Driver UUID
        format: uuid
        in: query
        name: driver_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.DriverStatsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Driver statistics
      tags:
      - stats
  /stats/team:
    get:
      parameters:
      - description: Team UUID
        format: uuid
        in: query
        name: team_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.TeamStatsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Team statistics
      tags:
      - stats
  /sudoku/drivers:
    get:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.MatrixDrivers'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get driver sudoku matrix
      tags:
      - sudoku
  /sudoku/teams:
    get:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.MatrixTeams'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get team sudoku matrix
      tags:
      - sudoku
...
  /users/favourite-driver:
    post:
      consumes:
      - application/json
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Driver payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Driver'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Toggle favourite driver
      tags:
      - users
  /users/favourite-team:
    post:
      consumes:
      - application/json
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Team payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Team'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Toggle favourite team
      tags:
      - users
swagger: "2.0"